Move exception classes to separate file
